CadmiumDEVS parser & no tags in SVG
When a simulation is loaded, the SVG IDs are used to create the dependence tree. Then, in the SimulationDEVS object, the method setDependencyNodes() adds the SVG nodes to the tree, so they can be used to draw the transitions. 
This version still needs a destination model in the transitions file, so I used the CadmiumDEVS parser to get the logs that I needed. 
I tested the code with my modified ABP model (with nested coupled models) and with the elevator FSM, and seems to work.
One last comment: I cover the case when multiple tags are used in the svg for the same port, using an array with the svg IDs in "simulation.json". We can remove that later if you think it's too much.